% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leebounds.R
\name{leebounds}
\alias{leebounds}
\title{Estimating Lee (2009) treatment effect bounds}
\usage{
leebounds(leedata)
}
\arguments{
\item{leedata}{data frame containing three fields
\item{res}{leedata$treat: binary treatment indicator}
\item{res}{leedata$selection: selection=1 if the outcome is observed }
\item{res}{leedata$outcome:   outcome=selection*outcome}}
}
\value{
A list containing the estimate of lower bound and upper bound
}
\description{
\code{leebounds} basic Lee (2009) bounds on treatment effect without covariates. Bounds are defined under monotonicity assumption
stating that treatment cannot hurt selection.
}
\examples{
n <- 500; x <- matrix(rnorm(n*5),nrow=n)
a <- runif(n); y <- a + rnorm(n,sd=.5)

ce.res <- ctseff(y,a,x, bw.seq=seq(.2,2,length.out=100))
plot.ctseff(ce.res)

# check that bandwidth choice is minimizer
plot(ce.res$bw.risk$bw,ce.res$bw.risk$risk)

}
\references{
David Lee (2009). Training, Wages, and Sample Selection: Estimating Sharp Bounds on Treatment Effects. \emph{The Review of Economic Studies, 76(3) 1071-1102}. \href{https://www.princeton.edu/~davidlee/wp/resrevision8.pdf}{https://www.princeton.edu/~davidlee/wp/resrevision8.pdf}
}
