% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leebounds.R
\name{leebounds_wout_monotonicity}
\alias{leebounds_wout_monotonicity}
\title{Compute Semenova (2019) bounds with conditional MTR assumption}
\usage{
leebounds_wout_monotonicity(leedata, s.hat)
}
\arguments{
\item{leedata, s.hat}{data frame with treat, selection, outcome; s.hat: predicted selection outcome}
}
\value{
Lee (2009) lower and upper bound
}
\description{
This function computes Semenova (2019) bounds upper and lower bound on the Average Treatment Effect under conditional MTR assumption. Its input argument
is a dataframe consisting of d=treat (binary treatment), s = selection (e.g., employment, test participation), and 
 outcome = sy observed only if s=1 (e.g., wage, test score). Lee (2009) bounds make two assumptions: (1) Treatment is randomly assigned and (2)
 There exists a partition of covariates X into X=X_0 + X_1 such that treatment helps selection if and only if x is in X_0 and 
 treatment hurts selection otherwise.
}
