% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortholeebounds.R
\name{ortho_bounds_ss_wt}
\alias{ortho_bounds_ss_wt}
\title{Compute orthogonal bounds (Semenova (2019))}
\usage{
ortho_bounds_ss_wt(
  leedata,
  treat_helps,
  y.hat,
  s.hat,
  weights = NULL,
  s_min = 0.1,
  ...
)
}
\arguments{
\item{leedata}{data frame with treat, selection, outcome}
}
\value{
lower and upper bound on ATE
}
\description{
This function computes orthogonal bounds Semenova (2019) upper and lower bound on the Average Treatment Effect. Its input argument
is a dataframe consisting of d=treat (binary treatment), s = selection (e.g., employment, test participation),  
 outcome = sy observed only if s=1 (e.g., wage, test score). In addition, it also requires the first-stage estimates
 of the conditional probability of selection and  the conditional quantile. It returns orthogonal (with respect to the first-stage)
 estimates of the upper and the lower bound. An optional vector of weights can be applied to the data to
 constructed weighted sample average.
}
