% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leebounds.R
\name{leebounds}
\alias{leebounds}
\title{Compute basic Lee (2009) bounds}
\usage{
leebounds(leedata)
}
\arguments{
\item{leedata}{data frame with treat, selection, outcome}
}
\value{
Lee (2009) lower and upper bound
}
\description{
This function computes basic Lee (2009) upper and lower bound on the Average Treatment Effect. Its input argument
is a dataframe consisting of d=treat (binary treatment), s = selection (e.g., employment, test participation), and 
 outcome = sy observed only if s=1 (e.g., wage, test score). Lee (2009) bounds make two assumptions: (1) Treatment is randomly assigned and (2)
 Treatment cannot hurt selection: S_1 >= S_0 a.s., where  S_1,S_0 are potential selection outcomes.
}
